backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.0
httpx==0.28.1
celery==5.4.0
redis==5.2.1
motor==3.7.0
aiofiles==24.1.0
python-multipart==0.0.20
loguru==0.7.3
slowapi==0.1.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
